% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_spec.R
\name{assert_spec}
\alias{assert_spec}
\title{Assert_spec
Function to check whether a dataset meets a desired specification}
\usage{
assert_spec(DAT, spec = NULL, file = NULL, error = FALSE)
}
\arguments{
\item{DAT}{Input dataset}

\item{spec}{YAML dataset specification}

\item{file}{file name containing YAML dataset specification}

\item{error}{Should the function error if issues are found ? (default is just a warning)}
}
\description{
Assert_spec
Function to check whether a dataset meets a desired specification
}
