% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAL.R
\docType{package}
\name{DAL-package}
\alias{DAL}
\alias{DAL-package}
\title{DAL: dataset validate}
\description{
This is a long descrption about how my package works it really needs to be updated.
This is some further description.
}
\author{
\strong{Maintainer}: Craig \email{cg@gmail.com}

}
