% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_spec.R
\name{assert_spec}
\alias{assert_spec}
\title{Assert_spec
Function to check whether a dataset meets a desired specification}
\usage{
assert_spec(DAT, spec = NULL, file = NULL, error = FALSE, warn = TRUE)
}
\arguments{
\item{DAT}{Input dataset}

\item{spec}{YAML dataset specification}

\item{file}{file name containing YAML dataset specification}

\item{error}{Should the function throw a error if issues are found ? (default False)}

\item{warn}{Should the function throw a warning if issues are found ? (default True)}
}
\description{
Assert_spec
Function to check whether a dataset meets a desired specification
}
